+++ PET/src/gui/tests/alternativeChoice/AlternativeChoicePanel.java
@@ -102,7 +102,7 @@ public class AlternativeChoicePanel extends javax.swing.JPanel implements ItemTe
     // End of variables declaration//GEN-END:variables
 
     @Override
-    public String saveItemTest(Long idTest) {
+    public String saveItemTest(Integer idTest) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
