+++ PET/src/gui/tests/select/SelectedPanel.java
@@ -142,7 +142,7 @@ public class SelectedPanel extends javax.swing.JPanel implements ItemTest{
     // End of variables declaration//GEN-END:variables
 
     @Override
-    public String saveItemTest(Long idTest) {
+    public String saveItemTest(Integer idTest) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
