+++ PET/src/gui/tests/sequencing/SequencingPanel.java
@@ -138,7 +138,7 @@ public class SequencingPanel extends javax.swing.JPanel implements ItemTest{
     // End of variables declaration//GEN-END:variables
 
     @Override
-    public String saveItemTest(Long idTest) {
+    public String saveItemTest(Integer idTest) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
